requirements.txt
matplotlib
pillow
numpy
opencv-python
imageio